Underlined enough of the important bits to get started.
In the words of an anonoymous and squirrly person, "Some blankity blank wrote these a year and a half a go and they haven't been maintained skince, so I skipped reading them, and just did it. It was a lot easier to do it that way."

Since I am feeling Le Tired, I'm going to skip reading the whole thing because that will make me le tireder.